Add product page

+++ app/product/[id]/AddToCartButton.tsx
@@ -1,15 +1,31 @@
 "use client";
 
-export default function AddToCartButton({ product }: { product: any }) {
+export default function AddToCartButton({
+  product,
+}: {
+  product: any;
+}) {
+  function addToCart() {
+    const cart = JSON.parse(localStorage.getItem("cart") || "[]");
+
+    const exists = cart.find((item: any) => item.id === product.id);
+
+    if (exists) {
+      alert("Товар уже в корзине");
+      return;
+    }
+
+    cart.push(product);
+
+    localStorage.setItem("cart", JSON.stringify(cart));
+
+    alert("Товар добавлен в корзину");
+  }
+
   return (
     <button
-      onClick={() => {
-        const cart = JSON.parse(localStorage.getItem("cart") || "[]");
-        cart.push(product);
-        localStorage.setItem("cart", JSON.stringify(cart));
-        window.location.href = "/cart";
-      }}
-      className="block text-center mt-6 w-full bg-yellow-400 text-black py-4 rounded-xl font-black"
+      onClick={addToCart}
+      className="w-full mt-8 bg-yellow-400 text-black py-5 rounded-2xl text-xl font-black hover:bg-yellow-300 hover:scale-[1.02] transition"
     >
       Добавить в корзину
     </button>
